Move header files from /glm to /include/glm subfolder
This makes it safer to include GLM as sources instead of install. By not
having the repository root as the header search path, but a dedicated
subfolder instead, chances of name clashes are reduced and the amount of
options displayed in IDEs is vastly reduced and disambiguated.

This change should not affect anyone relying on CMake's automatic
header search path setting of target dependencies. A re-run of cmake
may be necessary if it doesn't automatically pickup the changes. Already
installed/exported packages should not be affected.

Those already building from sources and managing search paths manually
need to switch from "path/to/glm" to "path/to/glm/include".